# backend/docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ISI-Eco-Report API",
        "description": "API certifiée de signalement et de gestion des déchets sauvages. Supporte PostgreSQL 17. Permet l'authentification (Sanctum), le signalement géolocalisé, l'affectation d'équipes, le suivi d'interventions et la gamification avec historique de points.",
        "version": "2.2.0"
    },
    "paths": {
...
        "/auth/logout": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Déconnexion",
                "operationId": "767c40d11a53a4b9ef07c97c6353be52",
                "responses": {
                    "204": {
                        "description": "Succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api",
                        "description": "Serveur API local"
                    }
                ]
            }
        },
...
        "/user": {
            "get": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Profil actuel",
                "operationId": "cc067d8fc4ee88652322a9816ca6079f",
                "responses": {
                    "200": {
                        "description": "Utilisateur connecté"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api",
                        "description": "Serveur API local"
                    }
                ]
            },
            "put": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Mettre à jour mon profil",
                "operationId": "65f39515cb7f09d4abf4e543173ce5cf",
                "responses": {
                    "200": {
                        "description": "Profil mis à jour"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api",
                        "description": "Serveur API local"
                    }
                ]
            }
        },
        "/users": {
            "get": {
                "tags": [
                    "Administration"
                ],
                "summary": "Lister les utilisateurs (ADMIN)",
                "operationId": "caed9c21fe975df9190788a75f4b758c",
                "responses": {
                    "200": {
                        "description": "Liste"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api",
                        "description": "Serveur API local"
                    }
                ]
            },
            "post": {
                "tags": [
                    "Administration"
                ],
                "summary": "Créer un utilisateur (ADMIN)",
                "operationId": "4f0b4b5cc1916208e42c21db68b4f0ae",
                "responses": {
                    "201": {
                        "description": "Créé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api",
                        "description": "Serveur API local"
                    }
                ]
            }
        },
        "/users/{id}": {
            "get": {
                "tags": [
                    "Administration"
                ],
                "summary": "Détails utilisateur (ADMIN)",
                "operationId": "28c8679c3bcc7f7e2469e732ce035974",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api",
                        "description": "Serveur API local"
                    }
                ]
            },
            "put": {
                "tags": [
                    "Administration"
                ],
                "summary": "Modifier utilisateur (ADMIN)",
                "operationId": "165a4ede0a0ecfc2bb8a11f6046ae366",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mis à jour"
                    },
                    "409": {
                        "description": "Conflit rôle/équipe"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api",
                        "description": "Serveur API local"
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Administration"
                ],
                "summary": "Supprimer utilisateur (ADMIN)",
                "operationId": "fe34178197b07119bd2b397f3b281e90",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Supprimé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api",
                        "description": "Serveur API local"
                    }
                ]
            }
        },
        "/types-dechets": {
            "get": {
                "tags": [
                    "Types Déchets"
                ],
                "summary": "Lister types déchets",
                "operationId": "81d461921ba08e277cdba92f2b421dd1",
                "responses": {
                    "200": {
                        "description": "Liste"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api",
                        "description": "Serveur API local"
                    }
                ]
            },
            "post": {
                "tags": [
                    "Administration"
                ],
                "summary": "Créer type déchet (ADMIN)",
                "operationId": "da2aacb24995403be3499cfb9a9c4d0e",
                "responses": {
                    "201": {
                        "description": "Créé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api",
                        "description": "Serveur API local"
                    }
                ]
            }
        },
        "/types-dechets/{id}": {
            "get": {
                "tags": [
                    "Types Déchets"
                ],
                "summary": "Détails type déchet",
                "operationId": "1a4d2b3dc98156ee1ad71bc6067e7ce5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api",
                        "description": "Serveur API local"
                    }
                ]
            },
            "put": {
                "tags": [
                    "Administration"
                ],
                "summary": "Modifier type déchet (ADMIN)",
                "operationId": "e4c35c0cff812efa9e5f519f67a4e334",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mis à jour"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api",
                        "description": "Serveur API local"
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Administration"
                ],
                "summary": "Supprimer type déchet (ADMIN)",
                "operationId": "365d1d55fae47948d2970f76ab22c2ed",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Supprimé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api",
                        "description": "Serveur API local"
                    }
                ]
            }
        },
        "/zones": {
            "get": {
                "tags": [
                    "Zones"
                ],
                "summary": "Lister les zones",
                "operationId": "a404b51a66ffd4c8b6607ce778ee734d",
                "responses": {
                    "200": {
                        "description": "Liste"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api",
                        "description": "Serveur API local"
                    }
                ]
            },
            "post": {
                "tags": [
                    "Administration"
                ],
                "summary": "Créer une zone (ADMIN)",
                "operationId": "21064b79efb997c3576162bb0f582e9e",
                "responses": {
                    "201": {
                        "description": "Créée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api",
                        "description": "Serveur API local"
                    }
                ]
            }
        },
        "/zones/{id}": {
            "get": {
                "tags": [
                    "Zones"
                ],
                "summary": "Détails zone",
                "operationId": "58e15fd4bd62362469f8a1e69b2fe441",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api",
                        "description": "Serveur API local"
                    }
                ]
            },
            "put": {
                "tags": [
                    "Administration"
                ],
                "summary": "Modifier zone (ADMIN)",
                "operationId": "adadb37c2955486617b51cead4ae93e7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mise à jour"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api",
                        "description": "Serveur API local"
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Administration"
                ],
                "summary": "Supprimer zone (ADMIN)",
                "operationId": "822e7cfe81089a777d23f6e37f3c6503",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Supprimée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api",
                        "description": "Serveur API local"
                    }
                ]
            }
        },
        "/equipes": {
            "get": {
                "tags": [
                    "Équipes"
                ],
                "summary": "Lister les équipes",
                "operationId": "d1b6aaa538683f55cf8bd02c714391e8",
                "responses": {
                    "200": {
                        "description": "Liste"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api",
                        "description": "Serveur API local"
                    }
                ]
            },
            "post": {
                "tags": [
                    "Administration"
                ],
                "summary": "Créer une équipe (ADMIN)",
                "operationId": "4a3073b804dfb5ed065b9173f76457fc",
                "responses": {
                    "201": {
                        "description": "Créée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api",
                        "description": "Serveur API local"
                    }
                ]
            }
        },
...
        "/signalements": {
            "get": {
                "tags": [
                    "Signalements"
                ],
                "summary": "Lister les signalements",
                "description": "Citoyen : voit les siens. Admin/Agent : voit tout.",
                "operationId": "49d2c9d81862854ebb2cdbc14914cc45",
                "responses": {
                    "200": {
                        "description": "Liste"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api",
                        "description": "Serveur API local"
                    }
                ]
            },
            "post": {
                "tags": [
                    "Signalements"
                ],
                "summary": "Créer un signalement",
                "operationId": "b0c4f8968fe5bd572a403ff3fba657e5",
                "responses": {
                    "201": {
                        "description": "Créé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api",
                        "description": "Serveur API local"
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Insérez votre jeton d'accès brut obtenu lors de la connexion.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}